chore: minimize diff
