+++ api/v1/cluster_funcs.go
@@ -39,7 +39,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/ptr"
 
 	"github.com/cloudnative-pg/cloudnative-pg/internal/configuration"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/system"
@@ -1294,93 +1293,6 @@ func (cluster *Cluster) GetSeccompProfile() *corev1.SeccompProfile {
 	return &corev1.SeccompProfile{
 		Type: corev1.SeccompProfileTypeRuntimeDefault,
 	}
-}
-
-// GetPodSecurityContext return the proper PodSecurityContext set in the cluster for Pods
-func (cluster *Cluster) GetPodSecurityContext() *corev1.PodSecurityContext {
-	// Under Openshift we inherit SecurityContext from the restricted security context constraint
-	if utils.HaveSecurityContextConstraints() {
-		return nil
-	}
-
-	uid := cluster.GetPostgresUID()
-	gid := cluster.GetPostgresGID()
-	defaultContext := &corev1.PodSecurityContext{
-		SeccompProfile: cluster.GetSeccompProfile(),
-		RunAsUser:      &uid,
-		RunAsGroup:     &gid,
-		RunAsNonRoot:   ptr.To(true),
-		FSGroup:        &gid,
-	}
-
-	if cluster.Spec.PodSecurityContext == nil {
-		return defaultContext
-	}
-
-	definedContext := cluster.Spec.PodSecurityContext
-
-	if definedContext.RunAsUser == nil {
-		definedContext.RunAsUser = defaultContext.RunAsUser
-	}
-	if definedContext.RunAsGroup == nil {
-		definedContext.RunAsGroup = defaultContext.RunAsGroup
-	}
-	if definedContext.SeccompProfile == nil {
-		definedContext.SeccompProfile = defaultContext.SeccompProfile
-	}
-
-	return definedContext
-}
-
-// GetSecurityContext return the proper SecurityContext in the cluster for Containers in Pods
-func (cluster *Cluster) GetSecurityContext() *corev1.SecurityContext {
-	trueValue := true
-	falseValue := false
-
-	defaultContext := &corev1.SecurityContext{
-		SeccompProfile: cluster.GetSeccompProfile(),
-		Capabilities: &corev1.Capabilities{
-			Drop: []corev1.Capability{
-				"ALL",
-			},
-		},
-		Privileged:               &falseValue,
-		RunAsNonRoot:             &trueValue,
-		ReadOnlyRootFilesystem:   &trueValue,
-		AllowPrivilegeEscalation: &falseValue,
-	}
-
-	if cluster.Spec.SecurityContext == nil {
-		return defaultContext
-	}
-
-	definedContext := cluster.Spec.SecurityContext
-	if definedContext.RunAsUser == nil {
-		definedContext.RunAsUser = defaultContext.RunAsUser
-	}
-	if definedContext.RunAsGroup == nil {
-		definedContext.RunAsGroup = defaultContext.RunAsGroup
-	}
-	if definedContext.SeccompProfile == nil {
-		definedContext.SeccompProfile = defaultContext.SeccompProfile
-	}
-	if definedContext.Capabilities == nil {
-		definedContext.Capabilities = defaultContext.Capabilities
-	}
-	if definedContext.Privileged == nil {
-		definedContext.Privileged = defaultContext.Privileged
-	}
-	if definedContext.RunAsNonRoot == nil {
-		definedContext.RunAsNonRoot = defaultContext.RunAsNonRoot
-	}
-	if definedContext.ReadOnlyRootFilesystem == nil {
-		definedContext.ReadOnlyRootFilesystem = defaultContext.ReadOnlyRootFilesystem
-	}
-	if definedContext.AllowPrivilegeEscalation == nil {
-		definedContext.AllowPrivilegeEscalation = defaultContext.AllowPrivilegeEscalation
-	}
-
-	return definedContext
 }
 
 // GetCoredumpFilter get the coredump filter value from the cluster annotation
