+++ pkg/specs/jobs.go
@@ -369,11 +369,11 @@ func CreatePrimaryJob(cluster apiv1.Cluster, nodeSerial int, role jobRole, initC
 							Command:         initCommand,
 							VolumeMounts:    CreatePostgresVolumeMounts(cluster),
 							Resources:       cluster.Spec.Resources,
-							SecurityContext: cluster.GetSecurityContext(),
+							SecurityContext: GetSecurityContext(&cluster),
 						},
 					},
 					Volumes:                   createPostgresVolumes(&cluster, instanceName),
-					SecurityContext:           cluster.GetPodSecurityContext(),
+					SecurityContext:           GetPodSecurityContext(&cluster),
 					Affinity:                  CreateAffinitySection(cluster.Name, cluster.Spec.Affinity),
 					Tolerations:               cluster.Spec.Affinity.Tolerations,
 					ServiceAccountName:        cluster.Name,
